Mybatis-plus 3.0.5

+++ renren-common/src/main/java/io/renren/common/utils/PageUtils.java
@@ -16,7 +16,7 @@
 
 package io.renren.common.utils;
 
-import com.baomidou.mybatisplus.plugins.Page;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 
 import java.io.Serializable;
 import java.util.List;
@@ -31,13 +31,13 @@ import java.util.List;
 public class PageUtils implements Serializable {
 	private static final long serialVersionUID = 1L;
 	//总记录数
-	private int totalCount;
+	private long totalCount;
 	//每页记录数
-	private int pageSize;
+	private long pageSize;
 	//总页数
-	private int totalPage;
+	private long totalPage;
 	//当前页数
-	private int currPage;
+	private long currPage;
 	//列表数据
 	private List<?> list;
 	
@@ -67,35 +67,35 @@ public class PageUtils implements Serializable {
 		this.totalPage = page.getPages();
 	}
 
-	public int getTotalCount() {
+	public long getTotalCount() {
 		return totalCount;
 	}
 
-	public void setTotalCount(int totalCount) {
+	public void setTotalCount(long totalCount) {
 		this.totalCount = totalCount;
 	}
 
-	public int getPageSize() {
+	public long getPageSize() {
 		return pageSize;
 	}
 
-	public void setPageSize(int pageSize) {
+	public void setPageSize(long pageSize) {
 		this.pageSize = pageSize;
 	}
 
-	public int getTotalPage() {
+	public long getTotalPage() {
 		return totalPage;
 	}
 
-	public void setTotalPage(int totalPage) {
+	public void setTotalPage(long totalPage) {
 		this.totalPage = totalPage;
 	}
 
-	public int getCurrPage() {
+	public long getCurrPage() {
 		return currPage;
 	}
 
-	public void setCurrPage(int currPage) {
+	public void setCurrPage(long currPage) {
 		this.currPage = currPage;
 	}
